Full pass over repo
Remove things we don't use, rename, simplify

Will move to apollographql/argocd-config-updater soon.
