+++ .github/workflows/ci.yml
@@ -42,23 +42,3 @@ jobs:
       - name: Test
         id: npm-ci-test
         run: npm run ci-test
-
-  # FIXME run the action against itself in CI
-  # test-action:
-  #   name: GitHub Actions Test
-  #   runs-on: ubuntu-latest
-
-  #   steps:
-  #     - name: Checkout
-  #       id: checkout
-  #       uses: actions/checkout@v4
-
-  #     - name: Test Local Action
-  #       id: test-action
-  #       uses: ./
-  #       with:
-  #         milliseconds: 1000
-
-  #     - name: Print Output
-  #       id: output
-  #       run: echo "${{ steps.test-action.outputs.time }}"
