+++ __tests__/update-git-refs.test.ts
@@ -1,20 +1,7 @@
-/**
- * Unit tests for the action's main functionality, src/main.ts
- *
- * These should be run as if the action was called from a workflow.
- * Specifically, the inputs listed in `action.yml` should be set as environment
- * variables following the pattern `INPUT_<INPUT_NAME>`.
- */
-
 import { readFile } from 'fs/promises';
 import { join } from 'path';
 import { GitHubClient } from '../src/github';
 import { updateGitRefs } from '../src/update-git-refs';
-
-// FIXME figure out logging (eg pass in a fake implementation of core
-// to code, or mock it, or capture process.stdout, etc)
-// Mock the GitHub Actions core library
-// let infoMock: jest.SpyInstance;
 
 const mockGitHubClient: GitHubClient = {
   async resolveRefToSha({ ref }) {
@@ -36,12 +23,6 @@ async function fixture(filename: string): Promise<string> {
 }
 
 describe('action', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    //infoMock = jest.spyOn(core, 'info').mockImplementation();
-  });
-
   it('updates git refs', async () => {
     const contents = await fixture('sample.yaml');
     const newContents = await updateGitRefs(contents, mockGitHubClient);
