+++ __tests__/update-promoted-values.test.ts
@@ -2,11 +2,6 @@ import { readFile } from 'fs/promises';
 import { join } from 'path';
 import { updatePromotedValues } from '../src/update-promoted-values';
 
-// FIXME figure out logging (eg pass in a fake implementation of core
-// to code, or mock it, or capture process.stdout, etc)
-// Mock the GitHub Actions core library
-// let infoMock: jest.SpyInstance;
-
 async function fixture(filename: string): Promise<string> {
   return await readFile(
     join(__dirname, '__fixtures__', 'update-promoted-values', filename),
@@ -15,12 +10,6 @@ async function fixture(filename: string): Promise<string> {
 }
 
 describe('action', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    //infoMock = jest.spyOn(core, 'info').mockImplementation();
-  });
-
   it('updates git refs', async () => {
     const contents = await fixture('sample.yaml');
     const newContents = await updatePromotedValues(contents, 'prod');
