+++ package.json
@@ -1,22 +1,14 @@
 {
-  "name": "typescript-action",
-  "description": "GitHub Actions TypeScript template",
+  "name": "argocd-config-updater",
+  "description": "ArgoCD Config Updater",
   "version": "0.0.0",
   "author": "",
   "private": true,
-  "homepage": "https://github.com/actions/typescript-action",
+  "homepage": "https://github.com/apollographql/argocd-config-updater",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/actions/typescript-action.git"
+    "url": "git+https://github.com/apollographql/argocd-config-updater.git"
   },
-  "bugs": {
-    "url": "https://github.com/actions/typescript-action/issues"
-  },
-  "keywords": [
-    "actions",
-    "node",
-    "setup"
-  ],
   "exports": {
     ".": "./dist/index.js"
   },
@@ -24,16 +16,15 @@
     "node": ">=20"
   },
   "scripts": {
-    "bundle": "npm run format:write && npm run copy-wasm && npm run package",
+    "bundle": "npm run format:write && rm -rf dist && mkdir dist && npm run copy-wasm && npm run package",
     "ci-test": "jest",
-    "coverage": "make-coverage-badge --output-path ./badges/coverage.svg",
     "format:write": "prettier --write **/*.ts",
     "format:check": "prettier --check **/*.ts",
-    "lint": "npx eslint . -c ./.github/linters/.eslintrc.yml",
+    "lint": "npx eslint . -c .eslintrc.yml",
     "package": "ncc build src/index.ts --license licenses.txt",
     "package:watch": "npm run package -- --watch",
     "test": "jest",
-    "all": "npm run format:write && npm run lint && npm run test && npm run coverage && npm run copy-wasm && npm run package",
+    "all": "npm run format:write && npm run lint && npm run test && npm run copy-wasm && npm run package",
     "copy-wasm": "cp node_modules/re2-wasm/build/wasm/re2.wasm dist/re2.wasm"
   },
   "license": "MIT",
@@ -55,22 +46,14 @@
     ],
     "transform": {
       "^.+\\.ts$": "ts-jest"
-    },
-    "coverageReporters": [
-      "json-summary",
-      "text",
-      "lcov"
-    ],
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "./src/**"
-    ]
+    }
   },
   "dependencies": {
     "@actions/core": "^1.10.1",
     "@actions/github": "^6.0.0",
     "@actions/glob": "^0.4.0",
     "@google-cloud/artifact-registry": "^3.1.0",
+    "@octokit/plugin-throttling": "^8.1.3",
     "lodash": "^4.17.21",
     "re2-wasm": "^1.0.2",
     "yaml": "^2.3.4"
@@ -88,8 +71,6 @@
     "eslint-plugin-jsonc": "^2.10.0",
     "eslint-plugin-prettier": "^5.0.1",
     "jest": "^29.7.0",
-    "js-yaml": "^4.1.0",
-    "make-coverage-badge": "^1.2.0",
     "prettier": "^3.0.3",
     "prettier-eslint": "^16.1.2",
     "ts-jest": "^29.1.1",
